+++ src/components/componentUserStatus/index.ts
@@ -1,0 +1,5 @@
+import { register } from '../../shared/functions';
+import ComponentUserStatus from './componentUserStatus';
+import '../componentSlider/index';
+
+register('component-user-status', ComponentUserStatus);