+++ src/components/pageHome/index.ts
@@ -3,6 +3,7 @@ import PageHome from './pageHome';
 import { register } from '../../shared/functions';
 
 import '../componentProgressBar';
+import '../componentUserStatus';
 import '../componentEmoji/index';
 import '../componentSearchResult/index';